// frontend/TradeTrove-GUI/src/app/models/stock-over-view.ts
export interface StockOverView {
  Symbol: string;
  AssetType: string;
  Name: string;
  Description: string;
  CIK: string;
  Exchange: string;
  Currency: string;
  Country: string;
  Sector: string;
  Industry: string;
  Address: string;
  OfficialSite: string;
  FiscalYearEnd: string;
  LatestQuarter: string;
  MarketCapitalization: number;
  EBITDA: number;
  PERatio: number;
  PEGRatio: number;
  BookValue: number;
  DividendPerShare: number;
  DividendYield: number;
  EPS: number;
  RevenuePerShareTTM: number;
  ProfitMargin: number;
  OperatingMarginTTM: number;
  ReturnOnAssetsTTM: number;
  ReturnOnEquityTTM: number;
  RevenueTTM: number;
  GrossProfitTTM: number;
  DilutedEPSTTM: number;
  QuarterlyEarningsGrowthYOY: number;
  QuarterlyRevenueGrowthYOY: number;
  AnalystTargetPrice: number;
  AnalystRatingStrongBuy: number;
  AnalystRatingBuy: number;
  AnalystRatingHold: number;
  AnalystRatingSell: number;
  AnalystRatingStrongSell: number;
  TrailingPE: number;
  ForwardPE: number;
  PriceToSalesRatioTTM: number;
  PriceToBookRatio: number;
  EVToRevenue: number;
  EVToEBITDA: number;
  Beta: number;
  "52WeekHigh": number;
  "52WeekLow": number;
  "50DayMovingAverage": number;
  "200DayMovingAverage": number;
  SharesOutstanding: number;
  SharesFloat: number;
  PercentInsiders: number;
  PercentInstitutions: number;
  DividendDate: string;
  ExDividendDate: string;

}